feat(introductory): add introductory section
The introductory contains greetings, name, and job title. The social
links which are the linkedin and github that should be in the header
to let the user easily connect to me. Also, the one important is the
link to resume that is currently work in progress. The link directs
the user to another domain.